update progpow quai 0.9.3

+++ sources/algo/progpow/firopow.hpp
@@ -13,7 +13,7 @@ namespace algo
         // Override Ethash
         constexpr uint32_t DAG_ITEM_PARENTS { 512u };
         constexpr uint32_t EPOCH_LENGTH{ 1300u };
-        constexpr uint32_t DAG_INIT_SIZE{ (1U << 30U) + (1U << 29U) };
+        constexpr uint64_t DAG_INIT_SIZE{ (1U << 30U) + (1U << 29U) };
         constexpr uint64_t DAG_COUNT_ITEMS_INIT{ DAG_INIT_SIZE / LEN_HASH_1024 };
 
         // Override ProgPOW
